Module libflate::finish

Finish and related types.

Structs

AutoFinish

A wrapper struct that completes the processing of the underlying instance when drops.

AutoFinishUnchecked

A wrapper struct that completes the processing of the underlying instance when drops.

Finish

Finish is a type that represents a value which may have an error occurred during the computation.

Traits

Complete

This trait allows to complete an I/O related processing.