Struct libflate::finish::AutoFinishUnchecked

pub struct AutoFinishUnchecked<T: Complete> { /* fields omitted */ }

A wrapper struct that completes the processing of the underlying instance when drops.

This calls Complete:::complete method of T when drops.

Note that this ignores the result of the invocation of Complete::complete(T).

Methods

impl<T: Complete> AutoFinishUnchecked<T>

pub fn new(inner: T) -> Self

Makes a new AutoFinishUnchecked instance.

Examples

use std::io;
use libflate::finish::AutoFinishUnchecked;
use libflate::gzip::Encoder;

let plain = b"Hello World!";
let mut buf = Vec::new();
let mut encoder = AutoFinishUnchecked::new(Encoder::new(&mut buf).unwrap());
io::copy(&mut &plain[..], &mut encoder).unwrap();

pub fn into_inner(self) -> T

Unwraps this AutoFinishUnchecked instance, returning the underlying instance.