Struct libflate::gzip::Encoder

pub struct Encoder<W, E = DefaultLz77Encoder> {
    header: Header,
    crc32: Crc32,
    input_size: u32,
    writer: Encoder<W, E>,
}

GZIP encoder.

Fields
header: Headercrc32: Crc32input_size: u32writer: Encoder<W, E>
Implementations

impl<W> Encoder<W, DefaultLz77Encoder>where W: Write,

pub fn new(inner: W) -> Result<Self>

Makes a new encoder instance.

Encoded GZIP stream is written to inner.

Examples

use std::io::Write;
use libflate::gzip::Encoder;

let mut encoder = Encoder::new(Vec::new()).unwrap();
encoder.write_all(b"Hello World!").unwrap();
encoder.finish().into_result().unwrap();

impl<W, E> Encoder<W, E>where W: Write, E: Lz77Encode,

pub fn with_options(inner: W, options: EncodeOptions<E>) -> Result<Self>

Makes a new encoder instance with specified options.

Encoded GZIP stream is written to inner.

Examples

use std::io::Write;
use libflate::gzip::{Encoder, EncodeOptions, HeaderBuilder};

let header = HeaderBuilder::new().modification_time(123).finish();
let options = EncodeOptions::new().no_compression().header(header);
let mut encoder = Encoder::with_options(Vec::new(), options).unwrap();
encoder.write_all(b"Hello World!").unwrap();

assert_eq!(encoder.finish().into_result().unwrap(),
           &[31, 139, 8, 0, 123, 0, 0, 0, 0, 3, 1, 12, 0, 243, 255, 72, 101, 108, 108,
             111, 32, 87, 111, 114, 108, 100, 33, 163, 28, 41, 28, 12, 0, 0, 0][..]);

pub fn header(&self) -> &Header

Returns the header of the GZIP stream.

Examples

use libflate::gzip::{Encoder, Os};

let encoder = Encoder::new(Vec::new()).unwrap();
assert_eq!(encoder.header().os(), Os::Unix);

pub fn finish(self) -> Finish<W, Error>

Writes the GZIP trailer and returns the inner stream.

Examples

use std::io::Write;
use libflate::gzip::Encoder;

let mut encoder = Encoder::new(Vec::new()).unwrap();
encoder.write_all(b"Hello World!").unwrap();

assert!(encoder.finish().as_result().is_ok())

Note

If you are not concerned the result of this encoding, it may be convenient to use AutoFinishUnchecked instead of the explicit invocation of this method.

use std::io;
use libflate::finish::AutoFinishUnchecked;
use libflate::gzip::Encoder;

let plain = b"Hello World!";
let mut buf = Vec::new();
let mut encoder = AutoFinishUnchecked::new(Encoder::new(&mut buf).unwrap());
io::copy(&mut &plain[..], &mut encoder).unwrap();

pub fn as_inner_ref(&self) -> &W

Returns the immutable reference to the inner stream.

pub fn as_inner_mut(&mut self) -> &mut W

Returns the mutable reference to the inner stream.

pub fn into_inner(self) -> W

Unwraps the Encoder, returning the inner stream.