Struct libflate::lz77::default::DefaultLz77Encoder

pub struct DefaultLz77Encoder {
    window_size: u16,
    buf: Vec<u8>,
}

A Lz77Encode implementation used by default.

Fields
window_size: u16buf: Vec<u8>
Implementations

impl DefaultLz77Encoder

pub fn new() -> Self

Makes a new encoder instance.

Examples

use libflate::deflate;
use libflate::lz77::{self, Lz77Encode, DefaultLz77Encoder};

let lz77 = DefaultLz77Encoder::new();
assert_eq!(lz77.window_size(), lz77::MAX_WINDOW_SIZE);

let options = deflate::EncodeOptions::with_lz77(lz77);
let _deflate = deflate::Encoder::with_options(Vec::new(), options);

pub fn with_window_size(size: u16) -> Self

Makes a new encoder instance with specified window size.

Larger window size is prefered to raise compression ratio, but it may require more working memory to encode and decode data.

Examples

use libflate::deflate;
use libflate::lz77::{self, Lz77Encode, DefaultLz77Encoder};

let lz77 = DefaultLz77Encoder::with_window_size(1024);
assert_eq!(lz77.window_size(), 1024);

let options = deflate::EncodeOptions::with_lz77(lz77);
let _deflate = deflate::Encoder::with_options(Vec::new(), options);