Module libflate::non_blocking

Implementations that can handle non-blocking I/O.

The implementations in this module can handle non-blocking Readers and Writers which will return ErrorKind::WouldBlock error when I/O operations would block.

If inner Readers and Writers return ErrorKind::WouldBlock error, Decoders and Encoders in this module will also return ErrorKind::WouldBlock.

If retrying the operation after the inner I/O become available, it will proceed successfully.

NOTICE

There is some performance penalty for non-blocking implementations against those that do not consider nonblocking I / O. So, it is recommended to use the latter if you are not need to handle non-blocking I/O.

Modules

deflate
The decoder of the DEFLATE format and algorithm.

gzip
The encoder and decoder of the GZIP format.

zlib
The encoder and decoder of the ZLIB format.